API Reference

This section provides a detailed reference for all functions and types defined in the CherenkovMediumBase module.

CherenkovMediumBase.DIPPR105Params

CherenkovMediumBase.DIPPR105

CherenkovMediumBase._hg_scattering_func

CherenkovMediumBase._refractive_index_fry

CherenkovMediumBase._sca_len_part_conc

CherenkovMediumBase.absorption_length

CherenkovMediumBase.calc_quan_fry_params

CherenkovMediumBase.cherenkov_angle

CherenkovMediumBase.dispersion

CherenkovMediumBase.dispersion_fry

CherenkovMediumBase.es_scattering

CherenkovMediumBase.es_scattering_cumulative

CherenkovMediumBase.es_scattering_integral

CherenkovMediumBase.group_refractive_index

CherenkovMediumBase.group_velocity

CherenkovMediumBase.make_inverse_es_polynomial

CherenkovMediumBase.material_density

CherenkovMediumBase.mixed_hg_generic_scattering_func

CherenkovMediumBase.mixed_hg_sl_scattering_func

CherenkovMediumBase.mixed_hg_sl_scattering_func_ppc

CherenkovMediumBase.phase_refractive_index

CherenkovMediumBase.pressure

CherenkovMediumBase.radiation_length

CherenkovMediumBase.sample_scattering_function

CherenkovMediumBase.scattering_length

CherenkovMediumBase.sl_scattering_func

CherenkovMediumBase.temperature

CherenkovMediumBase.DIPPR105Params — Type

DIPPR105Params

Parameters for the DIPPR105 formula

CherenkovMediumBase.DIPPR105 — Function

DIPPR105(temperature::Real, params::DIPPR105Params=DDBDIPR105Params)

Use DPPIR105 formula to calculate water density as function of temperature. temperature in K.

Reference: http://ddbonline.ddbst.de/DIPPR105DensityCalculation/DIPPR105CalculationCGI.exe?component=Water

Returns density in kg/m^3

CherenkovMediumBase._hg_scattering_func — Method

_hg_scattering_func(g::Real)

Henyey-Greenstein scattering in one plane.

Arguments

g::Real: mean scattering angle

Returns

typeof(g) cosine of a scattering angle sampled from the distribution

CherenkovMediumBase._refractive_index_fry — Method

refractive_index_fry(wavelength, salinity, temperature, pressure)

The phase refractive index of sea water according to a model from Quan & Fry.

wavelength is given in nm, salinity in permille, temperature in °C and pressure in atm

The original model is taken from: X. Quan, E.S. Fry, Appl. Opt., 34, 18 (1995) 3477-3480.

An additional term describing pressure dependence was included according to: Wolfgang H.W.A. Schuster, "Measurement of the Optical Properties of the Deep Mediterranean - the ANTARES Detector Medium.", PhD thesis (2002), St. Catherine's College, Oxford downloaded Jan 2011 from: http://www.physics.ox.ac.uk/Users/schuster/thesis0098mmjhuyynh/thesis.ps

Adapted from clsim (https://github.com/claudiok/clsim)

CherenkovMediumBase._sca_len_part_conc — Method

sca_len_part_conc(wavelength; vol_conc_small_part, vol_conc_large_part)

Calculates the scattering length (in m) for a given wavelength based on concentrations of small (vol_conc_small_part) and large (vol_conc_large_part) particles. wavelength is given in nm, volconcsmallpart and volconclargepart in ppm

C.D. Mobley "Light and Water", ISBN 0-12-502750-8, pag. 119.

CherenkovMediumBase.absorption_length — Method

absorption_length(medium::MediumProperties, wavelength)

Return absorption length at wavelength in units m.

wavelength is expected to be in units nm.

CherenkovMediumBase.calc_quan_fry_params — Method

calc_quan_fry_params(salinity::Real, temperature::Real, pressure::Real)

Helper function to get the parameters for the Quan & Fry formula as function of salinity, temperature and pressure.

CherenkovMediumBase.cherenkov_angle — Method

cherenkov_angle(medium, wavelength)

Calculate the cherenkov angle (in rad) for wavelength.

wavelength is expected to be in units nm.

CherenkovMediumBase.dispersion — Method

dispersion(medium, wavelength)

Return the dispersion dn/dλ at wavelength in units 1/nm.

wavelength is expected to be in units nm.

CherenkovMediumBase.dispersion_fry — Method

dispersion_fry(
    wavelength::T;
    salinity::Real,
    temperature::Real,
    pressure::Real) where {T <: Real}

Calculate the dispersion (dn/dλ) for the Quan & Fry model. Wavelength is given in nm, salinity in permille, temperature in °C and pressure in atm

CherenkovMediumBase.es_scattering — Method

es_scattering(cos_theta::T, b::T) where {T<:Real}

Einstein-Smoluchowsky Scattering PDF.

CherenkovMediumBase.es_scattering_cumulative — Method

es_scattering_integral(cos_theta::T, b::T) where {T<:Real}

Integral of ES scattering function from -1 to cos_theta

CherenkovMediumBase.es_scattering_integral — Method

es_scattering_integral(cos_theta::T, b::T) where {T<:Real}

Anti-derivative of the ES scattering function.

CherenkovMediumBase.group_refractive_index — Method

group_refractive_index(medium, wavelength)

Return the group refractive index at wavelength.

wavelength is expected to be in units nm.

CherenkovMediumBase.group_velocity — Method

group_velocity(medium, wavelength)

Return the group_velocity in m/ns at wavelength.

wavelength is expected to be in units nm.

CherenkovMediumBase.make_inverse_es_polynomial — Method

make_inverse_es_polynomial(b::T) where {T<:Real}

Make a 3rd order polynomial that approximates the inverse of the ES scattering function.

CherenkovMediumBase.material_density — Method

material_density(::MediumProperties)

This function returns the material density for a given medium.

CherenkovMediumBase.mixed_hg_generic_scattering_func — Method

mixed_hg_es_scattering_func(g::Real, hg_fraction::Real, poly_coeffs::NTuple{4, Real})

Mixture model of HG and generic scattering function.

CherenkovMediumBase.mixed_hg_sl_scattering_func — Method

mixed_hg_sl_scattering_func(g, hg_fraction)

Mixture model of HG and SL.

Arguments

g::Real: mean scattering angle

hg_fraction::Real: mixture weight of the HG component

CherenkovMediumBase.mixed_hg_sl_scattering_func_ppc — Method

mixed_hg_sl_scattering_func_ppc(g, hg_fraction)

Mixture model of HG and SL as implemented in PPC

CherenkovMediumBase.phase_refractive_index — Method

phase_refractive_index(medium, wavelength)

Return the phase refractive index at wavelength.

wavelength is expected to be in units nm.

CherenkovMediumBase.pressure — Method

pressure(::MediumProperties)

This function returns the pressure for a given medium.

CherenkovMediumBase.radiation_length — Method

radiation_length(medium::MediumProperties)

This function returns the radiation length for a given medium.

CherenkovMediumBase.sample_scattering_function — Method

scattering_function(medium::MediumProperties)

Return a scattering angle sampled from the scattering function of the medium.

CherenkovMediumBase.scattering_length — Method

scattering_length(medium::MediumProperties, wavelength)

Return scattering length at wavelength in units m.

wavelength is expected to be in units nm. Returned length is in units m.

CherenkovMediumBase.sl_scattering_func — Method

sl_scattering_func(g::Real)

Simplified-Liu scattering angle function. Implementation from: https://user-web.icecube.wisc.edu/~dima/work/WISC/ppc/spice/new/paper/a.pdf

Arguments

g::Real: mean scattering angle

CherenkovMediumBase.temperature — Method

temperature(::MediumProperties)

This function returns the temperature for a given medium.