API Reference

This section provides a detailed reference for all functions and types defined in the CherenkovMediumBase module.

CherenkovMediumBase.DIPPR105Params

CherenkovMediumBase.HenyeyGreenStein

CherenkovMediumBase.KopelevichScatteringModel

CherenkovMediumBase.PolynomialScatteringFunction

CherenkovMediumBase.QuanFryDispersion

CherenkovMediumBase.QuanFryDispersion

CherenkovMediumBase.QuanFryDispersion

CherenkovMediumBase.SimplifiedLiu

CherenkovMediumBase.TwoComponentScatteringFunction

CherenkovMediumBase.WavelengthIndependentAbsorptionModel

CherenkovMediumBase.WavelengthIndependentScatteringModel

CherenkovMediumBase.DIPPR105

CherenkovMediumBase._hg_scattering_func

CherenkovMediumBase._refractive_index_fry

CherenkovMediumBase._sca_len_part_conc

CherenkovMediumBase.absorption_length

CherenkovMediumBase.calc_quan_fry_params

CherenkovMediumBase.cherenkov_angle

CherenkovMediumBase.dispersion

CherenkovMediumBase.dispersion

CherenkovMediumBase.dispersion

CherenkovMediumBase.dispersion_fry

CherenkovMediumBase.es_scattering

CherenkovMediumBase.es_scattering_cumulative

CherenkovMediumBase.es_scattering_integral

CherenkovMediumBase.get_absorption_model

CherenkovMediumBase.get_dispersion_model

CherenkovMediumBase.get_scattering_model

CherenkovMediumBase.group_refractive_index

CherenkovMediumBase.group_velocity

CherenkovMediumBase.make_inverse_es_polynomial

CherenkovMediumBase.material_density

CherenkovMediumBase.phase_refractive_index

CherenkovMediumBase.phase_refractive_index

CherenkovMediumBase.phase_refractive_index

CherenkovMediumBase.pressure

CherenkovMediumBase.radiation_length

CherenkovMediumBase.sample_scattering_function

CherenkovMediumBase.scattering_length

CherenkovMediumBase.scattering_length

CherenkovMediumBase.sl_scattering_func

CherenkovMediumBase.temperature

CherenkovMediumBase.DIPPR105Params — Type

DIPPR105Params

Parameters for the DIPPR105 formula

CherenkovMediumBase.HenyeyGreenStein — Type

HenyeyGreenStein{T}

Struct for Henyey-Greenstein scattering function.

CherenkovMediumBase.KopelevichScatteringModel — Type

KopelevichScatteringModel{T, F<:AbstractScatteringFunction}

Struct for Kopelevich scattering model.

CherenkovMediumBase.PolynomialScatteringFunction — Type

PolynomialScatteringFunction{T, P <: ImmutablePolynomial{T}}

Struct for polynomial scattering function.

CherenkovMediumBase.QuanFryDispersion — Type

QuanFryDispersion{T <: Real}

Struct to hold parameters for the Quan & Fry dispersion model.

CherenkovMediumBase.QuanFryDispersion — Method

QuanFryDispersion(a01, a2, a3, a4)

Constructor for QuanFryDispersion with given parameters.

CherenkovMediumBase.QuanFryDispersion — Method

QuanFryDispersion(salinity, temperature, pressure)

Constructor for QuanFryDispersion using salinity, temperature, and pressure.

CherenkovMediumBase.SimplifiedLiu — Type

SimplifiedLiu{T}

Struct for Simplified-Liu scattering function.

CherenkovMediumBase.TwoComponentScatteringFunction — Type

TwoComponentScatteringFunction{F1<:AbstractScatteringFunction, F2<:AbstractScatteringFunction}

Struct for two-component mixture scattering model.

CherenkovMediumBase.WavelengthIndependentAbsorptionModel — Type

WavelengthIndependentAbsorptionModel{T} <: AbstractAbsorptionModel

Struct for wavelength-independent absorption model.

CherenkovMediumBase.WavelengthIndependentScatteringModel — Type

WavelengthIndependentScatteringModel{T, F<:AbstractScatteringFunction} <: AbstractScatteringModel

Struct for wavelength-independent scattering model.

CherenkovMediumBase.DIPPR105 — Function

DIPPR105(temperature::Real, params::DIPPR105Params=DDBDIPR105Params)

Use DPPIR105 formula to calculate water density as function of temperature. temperature in K.

Reference: http://ddbonline.ddbst.de/DIPPR105DensityCalculation/DIPPR105CalculationCGI.exe?component=Water

Returns density in kg/m^3

CherenkovMediumBase._hg_scattering_func — Method

_hg_scattering_func(g::Real)

Henyey-Greenstein scattering in one plane.

Arguments

g::Real: mean scattering angle

Returns

typeof(g) cosine of a scattering angle sampled from the distribution

CherenkovMediumBase._refractive_index_fry — Method

_refractive_index_fry(wavelength, quan_fry_params)

The phase refractive index of sea water according to a model from Quan & Fry.

wavelength is given in nm, salinity in permille, temperature in °C and pressure in atm

The original model is taken from: X. Quan, E.S. Fry, Appl. Opt., 34, 18 (1995) 3477-3480.

An additional term describing pressure dependence was included according to: Wolfgang H.W.A. Schuster, "Measurement of the Optical Properties of the Deep Mediterranean - the ANTARES Detector Medium.", PhD thesis (2002), St. Catherine's College, Oxford downloaded Jan 2011 from: http://www.physics.ox.ac.uk/Users/schuster/thesis0098mmjhuyynh/thesis.ps

Adapted from clsim (https://github.com/claudiok/clsim)

CherenkovMediumBase._sca_len_part_conc — Method

sca_len_part_conc(wavelength; vol_conc_small_part, vol_conc_large_part)

Calculates the scattering length (in m) for a given wavelength based on concentrations of small (vol_conc_small_part) and large (vol_conc_large_part) particles. wavelength is given in nm, volconcsmallpart and volconclargepart in ppm

C.D. Mobley "Light and Water", ISBN 0-12-502750-8, pag. 119.

CherenkovMediumBase.absorption_length — Method

absorption_length(medium::MediumProperties, wavelength)

Return absorption length at wavelength in units m. wavelength is expected to be in units nm.

CherenkovMediumBase.calc_quan_fry_params — Method

calc_quan_fry_params(salinity::Real, temperature::Real, pressure::Real)

Helper function to get the parameters for the Quan & Fry formula as function of salinity, temperature and pressure.

CherenkovMediumBase.cherenkov_angle — Method

cherenkov_angle(medium, wavelength)

Calculate the cherenkov angle (in rad) for wavelength.

wavelength is expected to be in units nm.

CherenkovMediumBase.dispersion — Method

dispersion(disperion_model::AbstractDispersionModel)

Calculate the dispersion for a given dispersion model.

CherenkovMediumBase.dispersion — Method

dispersion(medium::MediumProperties, wavelength)

Return the dispersion dn/dλ at wavelength in units 1/nm. wavelength is expected to be in units nm.

CherenkovMediumBase.dispersion — Method

dispersion(dispersion_model::QuanFryDispersion, wavelength)

Calculate the dispersion for the Quan & Fry dispersion model.

CherenkovMediumBase.dispersion_fry — Method

dispersion_fry(wavelength, a2, a3, a4)

Calculate the dispersion for the Quan & Fry dispersion model.

CherenkovMediumBase.es_scattering — Method

es_scattering(cos_theta::T, b::T) where {T<:Real}

Einstein-Smoluchowsky Scattering PDF.

CherenkovMediumBase.es_scattering_cumulative — Method

es_scattering_integral(cos_theta::T, b::T) where {T<:Real}

Integral of ES scattering function from -1 to cos_theta

CherenkovMediumBase.es_scattering_integral — Method

es_scattering_integral(cos_theta::T, b::T) where {T<:Real}

Anti-derivative of the ES scattering function.

CherenkovMediumBase.get_absorption_model — Method

get_absorption_model(medium::MediumProperties)

Return the absorption model for a given medium.

CherenkovMediumBase.get_dispersion_model — Method

get_dispersion_model(medium::MediumProperties)

Return the dispersion model for a given medium.

CherenkovMediumBase.get_scattering_model — Method

get_scattering_model(medium::MediumProperties)

Return the scattering model for a given medium.

CherenkovMediumBase.group_refractive_index — Method

group_refractive_index(medium, wavelength)

Return the group refractive index at wavelength.

wavelength is expected to be in units nm.

CherenkovMediumBase.group_velocity — Method

group_velocity(medium, wavelength)

Return the group_velocity in m/ns at wavelength.

wavelength is expected to be in units nm.

CherenkovMediumBase.make_inverse_es_polynomial — Method

make_inverse_es_polynomial(b::T) where {T<:Real}

Make a 3rd order polynomial that approximates the inverse of the ES scattering function.

CherenkovMediumBase.material_density — Method

material_density(medium::MediumProperties)

This function returns the material density for a given medium.

CherenkovMediumBase.phase_refractive_index — Method

phase_refractive_index(disperion_model::AbstractDispersionModel)

Calculate the phase refractive index for a given dispersion model.

CherenkovMediumBase.phase_refractive_index — Method

phase_refractive_index(medium::MediumProperties, wavelength)

Return the phase refractive index at wavelength. wavelength is expected to be in units nm.

CherenkovMediumBase.phase_refractive_index — Method

phase_refractive_index(disperion_model::QuanFryDispersion, wavelength)

Calculate the phase refractive index for the Quan & Fry dispersion model.

CherenkovMediumBase.pressure — Method

pressure(medium::MediumProperties)

This function returns the pressure for a given medium.

CherenkovMediumBase.radiation_length — Method

radiation_length(medium::MediumProperties)

This function returns the radiation length for a given medium.

CherenkovMediumBase.sample_scattering_function — Method

sample_scattering_function(medium::MediumProperties)

Return a cos(scattering angle) sampled from the scattering function of the medium.

CherenkovMediumBase.scattering_length — Method

scattering_length(model::KopelevichScatteringModel, wavelength::Real)

Calculate the scattering length for the Kopelevich scattering model.

CherenkovMediumBase.scattering_length — Method

scattering_length(medium::MediumProperties, wavelength)

Return scattering length at wavelength in units m. wavelength is expected to be in units nm. Returned length is in units m.

CherenkovMediumBase.sl_scattering_func — Method

sl_scattering_func(g::Real)

Simplified-Liu scattering angle function. Implementation from: https://user-web.icecube.wisc.edu/~dima/work/WISC/ppc/spice/new/paper/a.pdf

Arguments

g::Real: mean scattering angle

CherenkovMediumBase.temperature — Method

temperature(medium::MediumProperties)

This function returns the temperature for a given medium.